Move some files to LibHac.Common
