+++ src/LibHac/Common/ThrowHelper.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace LibHac;
+namespace LibHac.Common;
 
 internal static class ThrowHelper
 {
