+++ src/LibHac/Crypto/Impl/AesCtrModeNi.cs
@@ -4,6 +4,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.Intrinsics;
 using System.Runtime.Intrinsics.X86;
+using LibHac.Common;
 
 namespace LibHac.Crypto.Impl;
