+++ src/LibHac/Diag/Abort.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
+using LibHac.Common;
 
 namespace LibHac.Diag;
