+++ src/LibHac/FsSystem/Aes128CtrTransform.cs
@@ -3,6 +3,7 @@ using System.Buffers;
 using System.Buffers.Binary;
 using System.Runtime.InteropServices;
 using System.Security.Cryptography;
+using LibHac.Common;
 using LibHac.Util;
 
 namespace LibHac.FsSystem;
