+++ src/LibHac/FsSystem/Aes128XtsTransform.cs
@@ -28,6 +28,7 @@ using System;
 using System.Buffers;
 using System.Runtime.InteropServices;
 using System.Security.Cryptography;
+using LibHac.Common;
 using LibHac.Util;
 
 namespace LibHac.FsSystem;
