+++ src/LibHac/FsSystem/HierarchicalIntegrityVerificationStorage.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Security.Cryptography;
 using System.Text;
+using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Util;
