+++ src/LibHac/FsSystem/IntegrityVerificationStorage.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Buffers;
 using System.IO;
+using LibHac.Common;
 using LibHac.Crypto;
 using LibHac.Fs;
 using LibHac.Tools.FsSystem.Save;
