+++ src/LibHac/Npdm/ServiceAccessControl.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using LibHac.Common;
 
 namespace LibHac.Npdm;
