+++ src/LibHac/Tools/FsSystem/Save/AllocationTable.cs
@@ -2,6 +2,7 @@
 using System.Diagnostics;
 using System.IO;
 using System.Runtime.InteropServices;
+using LibHac.Common;
 using LibHac.Fs;
 using LibHac.FsSystem;
