+++ src/LibHac/Tools/FsSystem/Save/Header.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using LibHac.Common;
 using LibHac.Common.Keys;
 using LibHac.Crypto;
 using LibHac.Fs;
