+++ src/LibHac/Tools/Ro/Nro.cs
@@ -1,4 +1,5 @@
 ﻿using System.IO;
+using LibHac.Common;
 using LibHac.Fs;
 using LibHac.FsSystem;
