+++ src/LibHac/Tools/Settings/Calibration.cs
@@ -1,5 +1,6 @@
 ﻿using System.IO;
 using System.Text;
+using LibHac.Common;
 
 namespace LibHac.Tools.Settings;
