+++ src/hactoolnet/EnumStrings.cs
@@ -1,4 +1,4 @@
-﻿using LibHac;
+﻿using LibHac.Common;
 using LibHac.Fs;
 using LibHac.FsSystem;
 using LibHac.Ncm;
