+++ src/hactoolnet/Options.cs
@@ -1,4 +1,5 @@
 ﻿using LibHac;
+using LibHac.Common;
 using LibHac.Common.Keys;
 using LibHac.FsSystem;
