+++ tests/LibHac.Tests/DebugAssertHandler.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using LibHac.Common;
 
 namespace LibHac.Tests;
