+++ tests/LibHac.Tests/FsSystem/BucketTreeBuilderTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Runtime.CompilerServices;
+using LibHac.Common;
 using LibHac.Fs;
 using LibHac.FsSystem;
 using LibHac.Tests.Common;
